re-organise and update params
