+++ json-schema-generator/output/com/example/project/MyLlamaNumberOne.java
@@ -1,0 +1,8 @@
+import io.micronaut.serde.annotation.Serdeable;
+import java.lang.String;
+
+@Serdeable
+public record MyLlamaNumberOne(
+    String name
+) {
+}
