+++ json-schema-generator/output/com/example/project/TestRecord.java
@@ -1,0 +1,12 @@
+import io.micronaut.serde.annotation.Serdeable;
+import jakarta.validation.constraints.DecimalMin;
+import jakarta.validation.constraints.Size;
+import java.lang.Float;
+import java.util.List;
+import java.util.Set;
+
+@Serdeable
+public record TestRecord(
+    @Size(min = 1) List<@Size(min = 2) Set<@DecimalMin(10.0) Float>> arrayMulti
+) {
+}
