+++ json-schema-generator/src/main/java/io/micronaut/jsonschema/generator/aggregator/DefinitionsAggregator.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2017-2024 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.jsonschema.generator.aggregator;
+
+import io.micronaut.core.annotation.Internal;
+import jakarta.inject.Singleton;
+
+import java.util.HashMap;
+
+/**
+ * An aggregator for storing and accessing definitions from json schema.
+ */
+@Internal
+@Singleton
+public class DefinitionsAggregator
+{
+    private static HashMap<String, Object> definitions = new HashMap<>();
+
+}
